refactor: 온보딩 지역선택 그리드 별도의 컴포넌트로 분리
