+++ pages/onboard/location/index.page.tsx
@@ -3,18 +3,14 @@ import Link from "next/link";
 import { useTheme } from "@emotion/react";
 import useOnboardStore from "@/store/onboard";
 import { ROUTES } from "@/constants/routes";
-import { locations } from "@/mocks/mockData/location";
+import Location from "@/components/onboard/Location/Location";
 import Button from "@/components/shared/Button/Button";
 import Text from "@/components/shared/Text/Text";
 import * as S from "./location.styled";
 
 function OnboardLocationPage() {
   const { colors } = useTheme();
-  const { location, setLocation } = useOnboardStore();
-
-  const handleClick = (newLocation: string) => {
-    setLocation(newLocation);
-  };
+  const { location } = useOnboardStore();
 
   return (
     <S.Wrap>
@@ -37,25 +33,7 @@ function OnboardLocationPage() {
           선택한 지역(최대 2개)을 중심으로 스토어를 추천드려요.
         </Text>
       </S.TextContainer>
-      <S.LocationContainer>
-        {locations.map(locationItem => (
-          <button
-            key={locationItem.id}
-            className={location.includes(locationItem.name) ? "isSelected" : ""}
-            onClick={() => handleClick(locationItem.name)}
-          >
-            <Text weight={500} size={15}>
-              {locationItem.name}
-            </Text>
-            <Text weight={500} size={12}>{`(${locationItem.storeCounts}개)`}</Text>
-            {locationItem.label && (
-              <S.LabelTag type="single" label={locationItem.label}>
-                {locationItem.label}
-              </S.LabelTag>
-            )}
-          </button>
-        ))}
-      </S.LocationContainer>
+      <Location />
       <S.Footer>
         <Link href={ROUTES.ONBOARDING_CATEGORY_PAGE}>
           <Button disabled={!location.length} type="button" label="next" shape="square">
